basic+vm: Tweak script quoting and prereqs handling
The new onboarding `just get-started` target installs moss to a
`/usr/local` prefix by default.

All script path variables should now be quoted as well.

+++ basic-setup.sh
@@ -43,9 +43,9 @@ die() {
 }
 
 checkPrereqs () {
-    # prerequisite checks
+    printInfo "Checking prerequisites..."
     test -f ./pkglist-base || die "\nRun this script from the root of the img-tests/ repo clone!\n"
-    test -x /usr/bin/moss || die "\n${0} assumes moss is installed. See https://github.com/serpent-os/onboarding/\n"
+    test -x $(command -v moss) || die "\n${0} assumes moss is installed. See https://github.com/serpent-os/moss/\n"
 }
 
 # base packages
@@ -72,13 +72,16 @@ EOF
 }
 
 basicSetup () {
+    # NB: This will fail if moss is an alias!
+    local moss="$(command -v moss)"
+
     MSG="Removing old ${SOSROOT} directory..."
     printInfo "${MSG}"
     sudo rm -rf "${SOSROOT}" || die "${MSG}"
 
     MSG="Creating new ${SOSROOT} directory w/baselayout skeleton..."
     printInfo "${MSG}"
-    sudo mkdir -pv "${SOSROOT}"/{etc,proc,run,sys,var,var/local,${BOULDERCACHE}} || die "${MSG}"
+    sudo mkdir -pv "${SOSROOT}"/{etc,proc,run,sys,var,var/local,"${BOULDERCACHE}"} || die "${MSG}"
 
     # No longer necessary -- moss triggers have been fixed to respect trigger dep order now
     #MSG="Ensuring that we get a working nss-systemd-compatible nssswitch.conf..."
@@ -92,31 +95,31 @@ basicSetup () {
 
     MSG="Adding volatile serpent os repository..."
     printInfo "${MSG}"
-    sudo moss -D "${SOSROOT}" -y repo add volatile https://dev.serpentos.com/volatile/x86_64/stone.index -p0 || die "${MSG}"
+    sudo ${moss} -D "${SOSROOT}" -y repo add volatile https://dev.serpentos.com/volatile/x86_64/stone.index -p0 || die "${MSG}"
 
     MSG="Installing packages..."
     printInfo "${MSG}"
-    sudo moss -D "${SOSROOT}" -y --cache "${BOULDERCACHE}" install "${PACKAGES[@]}" || die "${MSG}"
+    sudo ${moss} -D "${SOSROOT}" -y --cache "${BOULDERCACHE}" install "${PACKAGES[@]}" || die "${MSG}"
 
     MSG="Setting up an empty root password by default..."
     printInfo "${MSG}"
     sudo chroot "${SOSROOT}" /usr/bin/passwd -d root
     sudo rm -vf issue
-    cp -v "${SOSROOT}"/etc/issue issue
+    test -f "${SOSROOT}"/etc/issue && cp -v "${SOSROOT}"/etc/issue issue
     echo -e "By default, the root user has no password.\n\nUse the passwd command to change it.\n" >> issue
-    sudo cp -v issue "${SOSROOT}"/etc/issue
+    sudo mv -v issue "${SOSROOT}"/etc/issue
 
     MSG="Preparing local-x86_64 profile directory..."
     printInfo "${MSG}"
-    sudo mkdir -pv ${SOSROOT}/${BOULDERCACHE}/repos/local-x86_64/ || die "${MSG}"
+    sudo mkdir -pv "${SOSROOT}/${BOULDERCACHE}/repos/local-x86_64/" || die "${MSG}"
 
     MSG="Creating a moss stone.index file for the local-x86_64 profile..."
     printInfo "${MSG}"
-    sudo moss -y index ${SOSROOT}/${BOULDERCACHE}/repos/local-x86_64/ || die "${MSG}"
+    sudo ${moss} -y index "${SOSROOT}/${BOULDERCACHE}/repos/local-x86_64/" || die "${MSG}"
 
     MSG="Adding local-x86_64 profile to list of active repositories..."
     printInfo "${MSG}"
-    sudo moss -D "${SOSROOT}" -y repo add local-x86_64 file://${BOULDERCACHE}/repos/local-x86_64/stone.index -p10 || die "${MSG}"
+    sudo ${moss} -D "${SOSROOT}" -y repo add local-x86_64 "file://${BOULDERCACHE}/repos/local-x86_64/stone.index" -p10 || die "${MSG}"
 }
 
 # clean up env
